Restructure folders to include controllers and rename files
add controllers folder and create blank files that will house controller files

correct all references to renamed files, ensure the server starts and tests pass
