+++ test/routes_test.js
@@ -1,6 +1,6 @@
 const chai = require('chai');
 const chaiHttp = require('chai-http');
-const app = require('../server/app').default;
+const app = require('../dist/app').default;
 
 const { assert } = chai;
 const { expect } = chai;
